move review request functions to a standalone file
